refactor: product card and move product data to const variable
