--- src/constants/favorite-products.js
@@ -1,20 +1,0 @@
-import productImage from "../assets/GLOD WEB/BRSFC 01.jpg";
-
-export const favoriteProducts = [
-  {
-    id: 1,
-    name: "Nama Barang",
-    color: "Light Grey",
-    normalPrice: 300000,
-    price: 150000,
-    mainImage: productImage,
-  },
-  {
-    id: 2,
-    name: "Nama Barang",
-    color: "Light Grey",
-    normalPrice: 300000,
-    price: 150000,
-    mainImage: productImage,
-  },
-];
