+++ src/pages/catalogue/index.js
@@ -6,109 +6,7 @@ import {
   ProductCard,
 } from "../../components";
 import { seo_meta } from "../../constants/seo";
-
-import BRSFC01 from "../../assets/GLOD WEB/BRSFC 01.jpg";
-import BRSFC02 from "../../assets/GLOD WEB/BRSFC 02.jpg";
-import BRSFC03 from "../../assets/GLOD WEB/BRSFC 03.jpg";
-import BRSFC04 from "../../assets/GLOD WEB/BRSFC 04.jpg";
-import BRSFC05 from "../../assets/GLOD WEB/BRSFC 05.jpg";
-import SHINDYO1 from "../../assets/GLOD WEB/SHINDY 00101.jpg";
-import SHINDYO2 from "../../assets/GLOD WEB/SHINDY 00102.jpg";
-import SHINDYO3 from "../../assets/GLOD WEB/SHINDY 00103.jpg";
-import SHINDYO4 from "../../assets/GLOD WEB/SHINDY 00201.jpg";
-import SHINDYO5 from "../../assets/GLOD WEB/SHINDY 00202.jpg";
-import SHINDYO6 from "../../assets/GLOD WEB/SHINDY 00203.jpg";
-
-const catalogueProducts = [
-  {
-    id: 1,
-    tags: ["BRSFC", "jersey"],
-    image: BRSFC01,
-    product: "BRSFC 01",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-  {
-    id: 2,
-    tags: ["BRSFC", "jersey"],
-    image: BRSFC02,
-    product: "BRSFC 02",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-  {
-    id: 3,
-    tags: ["BRSFC", "jersey"],
-    image: BRSFC03,
-    product: "BRSFC 03",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-  {
-    id: 4,
-    tags: ["BRSFC", "jersey"],
-    image: BRSFC04,
-    product: "BRSFC 04",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-  {
-    id: 5,
-    tags: ["BRSFC", "jersey"],
-    image: BRSFC05,
-    product: "BRSFC 05",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-  {
-    id: 6,
-    tags: ["SHINDY", "jersey"],
-    image: SHINDYO1,
-    product: "SHINDY 01",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-  {
-    id: 7,
-    tags: ["SHINDY", "jersey"],
-    image: SHINDYO2,
-    product: "SHINDY 02",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-  {
-    id: 8,
-    tags: ["SHINDY", "jersey"],
-    image: SHINDYO3,
-    product: "SHINDY 03",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-  {
-    id: 9,
-    tags: ["SHINDY", "jersey"],
-    image: SHINDYO4,
-    product: "SHINDY 04",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-  {
-    id: 10,
-    tags: ["SHINDY", "jersey"],
-    image: SHINDYO5,
-    product: "SHINDY 05",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-  {
-    id: 11,
-    tags: ["SHINDY", "jersey"],
-    image: SHINDYO6,
-    product: "SHINDY 06",
-    price: 50000,
-    colors: ["red", "black"],
-  },
-];
+import { catalogueProducts } from "../../constants/products"
 
 export default function CataloguePage(props) {
   const tagFilter = ["jersey"];
